[Refactor/#95] spring-doc yml(sort setting) add
