+++ src/main/java/com/example/waggle/global/security/oauth2/handler/OAuth2AuthenticationFailureHandler.java
@@ -1,8 +1,6 @@
 package com.example.waggle.global.security.oauth2.handler;
 
-import com.example.waggle.global.security.oauth2.cookie.CookieUtil;
 import jakarta.servlet.ServletException;
-import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
