+++ src/main/java/com/example/waggle/global/security/oauth2/handler/OAuth2AuthenticationSuccessHandler.java
@@ -1,32 +1,19 @@
 package com.example.waggle.global.security.oauth2.handler;
 
-import com.example.waggle.domain.member.entity.Role;
-import com.example.waggle.global.exception.handler.SecurityHandler;
-import com.example.waggle.global.payload.ApiResponseDto;
-import com.example.waggle.global.payload.code.ErrorStatus;
-import com.example.waggle.global.security.CustomUserDetails;
 import com.example.waggle.global.security.JwtToken;
 import com.example.waggle.global.security.TokenService;
-import com.example.waggle.global.security.oauth2.cookie.CookieUtil;
-import com.nimbusds.jose.shaded.gson.Gson;
+import com.example.waggle.global.util.CookieUtil;
 import jakarta.servlet.ServletException;
-import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import netscape.javascript.JSObject;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.http.*;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.web.authentication.SimpleUrlAuthenticationSuccessHandler;
 import org.springframework.stereotype.Component;
-import org.springframework.web.client.RestTemplate;
-import org.springframework.web.util.UriComponentsBuilder;
 
 import java.io.IOException;
-import java.net.URI;
-import java.util.Optional;
 
 
 @Slf4j
@@ -47,8 +34,8 @@ public class OAuth2AuthenticationSuccessHandler extends SimpleUrlAuthenticationS
             return;
         }
         String refreshToken = jwtToken.getRefreshToken();
-        CookieUtil.addCookie(response, "refresh_token", refreshToken,60);
-        CookieUtil.deleteCookie(request,response,"JSESSIONID");
+        CookieUtil.addCookie(response, "refresh_token", refreshToken, 60);
+        CookieUtil.deleteCookie(request, response, "JSESSIONID");
 
         getRedirectStrategy().sendRedirect(request, response, redirectUri);
     }
