+++ src/main/resources/application-real.yml
@@ -1,9 +1,9 @@
 spring:
-#  profiles:
-#    include:
-#      - security
-#      - real-db
-#      - s3
+  #  profiles:
+  #    include:
+  #      - security
+  #      - real-db
+  #      - s3
   jpa:
     properties:
       hibernate:
@@ -41,4 +41,7 @@ server:
     whitelabel.enabled: true
 
 springdoc:
+  swagger-ui:
+    tags-sorter: alpha
+    operations-sorter: method
   use-fqn: true
